phase02-todo-app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==25.1.0
python-multipart==0.0.20
python-dotenv==1.0.1
aiosqlite==0.20.0
alembic==1.13.3
psycopg2-binary
email-validator
google-generativeai==0.8.3
httpx==0.27.0
sqlalchemy==2.0.35